visionbot_prototype_0.5/requirements.txt
Flask==2.3.3
flask-cors>=4.0.0
ultralytics>=8.3.0
opencv-python>=4.8.0
numpy>=1.24.0
requests>=2.31.0
Pillow>=9.5.0
Werkzeug==2.3.7
python-dotenv>=1.0.0